+++ internal/tags/main.go
@@ -6,10 +6,34 @@
 package tags
 
 import (
+	"regexp"
 	"sort"
 	"strings"
 
 	"github.com/pkg/errors"
+)
+
+// JSONMetric defines an individual metric received in JSON
+type JSONMetric struct {
+	Tags  []string    `json:"_tags"`
+	Type  string      `json:"_type"`
+	Value interface{} `json:"_value"`
+}
+
+// JSONMetrics holds list of JSON metrics
+type JSONMetrics map[string]JSONMetric
+
+const (
+	// Delimiter defines character separating category from value in a tag e.g. location:london
+	Delimiter = ":"
+	// Separator defines character separating tags in a list e.g. os:centos,location:sfo
+	Separator       = ","
+	replacementChar = "_"
+)
+
+var (
+	valid   = regexp.MustCompile(`^[^:,]+:[^:,]+(,[^:,]+:[^:,]+)*$`)
+	cleaner = regexp.MustCompile(`[\[\]'"` + "`]")
 )
 
 // PrepStreamTags accepts a comma delimited list of key:value pairs
